requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlmodel==0.0.14
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9
asyncpg==0.29.0
aioredis==2.0.1
alembic==1.13.1
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
bcrypt==4.1.2